requirements.txt
streamlit==1.18.1
pillow==9.2.0